% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealista18.R
\docType{data}
\name{Barcelona_POIS}
\alias{Barcelona_POIS}
\title{Places of interest in Barcelona.}
\format{
A list with two objects:
\describe{
  \item{City_Center}{Coordinates of Barcelona's city center}
  \item{Diagonal}{Avinguda Diagonal: Coordinates of Diagonal Avenue)}
  }

An object of class \code{list} of length 3.
}
\source{
Idealista (https://www.idealista.com/)
}
\usage{
data(Barcelona_POIS)

Barcelona_POIS
}
\description{
Geospatial references for places of interest in Barcelona
}
\examples{
 library(idealista18)
 data(Barcelona_POIS)
 summary(Barcelona_POIS)
}
\keyword{datasets}
