% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealista18.R
\docType{data}
\name{Barcelona_Sale}
\alias{Barcelona_Sale}
\title{Sale prices of residential property in Barcelona in 2018.}
\format{
An object of class `sf` (inherits from `data.frame`) with 84,280 rows and 57 variables:
\describe{
  \item{ASSETID}{Unique identifier of the asset}
  \item{PERIOD}{Date YYYYMM, indicates the quarter when the ad was extracted we used YYYY03 for the 1st Q, YYYY06 the 2nd, YYYY09 for the 3rd and YYYY12 for the 4th}
  \item{PRICE}{Asking price for the ad at idealista expressed in euros}
  \item{UNITPRICE}{Price in euros per square meter}
  \item{ADTYPOLOGYID}{Asset Typology (home, chalet)}
  \item{ADOPERATIONID}{Asset Operation (rent, sales)}
  \item{CONSTRUCTEDAREA}{Home built area in sq.m}
  \item{ROOMNUMBER}{Number of bedrooms}
  \item{BATHNUMBER}{Number of bathrooms}
  \item{HASTERRACE}{Dummy variable for terrace (takes 1 if there is a terrace, 0 otherwise)}
  \item{HASLIFT}{Dummy variable for lift (takes 1 if there is a lift in the building, 0 otherwise)}
  \item{HASAIRCONDITIONING}{Dummy variable for AA (takes 1 if there is a AA, 0 otherwise)}
  \item{AMENITYID}{Indicates the amenities included (1 - no furniture, no kitchen amenities, 2 - kitchen amenities, no furniture, 3 - kitchen amenities, furniture)}
  \item{HASPARKINGSPACE}{Dummy variable for parking (takes 1 if parking is included in the Ad, 0 otherwise)}
  \item{ISPARKINGSPACEINCLUDEDINPRICE}{Dummy variable for parking (takes 1 if parking is included in the Ad, 0 otherwise)}
  \item{PARKINGSPACEPRICE}{Price of parking Space in euros}
  \item{HASNORTHORIENTATION}{Dummy variable for orientation (takes 1 if orientation is North in the Ad, 0 otherwise) - Important note: orientation features are not orthogonal features, a house oriented to the north can be also oriented to the east}
  \item{HASSOUTHORIENTATION}{Dummy variable for orientation (takes 1 if orientation is South in the Ad, 0 otherwise)  - Important note: orientation features are not orthogonal features, a house oriented to the north can be also oriented to the east}
  \item{HASEASTORIENTATION}{Dummy variable for orientation (takes 1 if orientation is East in the Ad, 0 otherwise)  - Important note: orientation features are not orthogonal features, a house oriented to the north can be also oriented to the east}
  \item{HASWESTORIENTATION}{Dummy variable for orientation (takes 1 if orientation is West in the Ad, 0 otherwise)  - Important note: orientation features are not orthogonal features, a house oriented to the north can be also oriented to the east}
  \item{HASBOXROOM}{Dummy variable for boxroom (takes 1 if boxroom is included in the Ad, 0 otherwise)}
  \item{HASWARDROBE}{Dummy variable for wardrobe (takes 1 if wardrobe is included in the Ad, 0 otherwise)}
  \item{HASSWIMMINGPOOL}{Dummy variable for swimming pool (takes 1 if swimming pool is included in the Ad, 0 otherwise)}
  \item{HASDOORMAN}{Dummy variable for doorman (takes 1 if there is a doorman in the building, 0 otherwise)}
  \item{HASGARDEN}{Dummy variable for garden (takes 1 if there is a garden in the building, 0 otherwise)}
  \item{ISDUPLEX}{Dummy variable for duplex (takes 1 if it is a duplex, 0 otherwise)}
  \item{ISSTUDIO}{Dummy variable for bachelor apartment (studio in Spanish) (takes 1 if it is a single-person apartment, 0 otherwise)}
  \item{ISINTOPFLOOR}{Dummy variable indicating if the apartment is located in the top floor (takes 1 on the top floor 0 otherwise)}
  \item{CONSTRUCTIONYEAR}{Construction year (source: advertiser)}
  \item{FLOORCLEAN}{Indicates flat floornumber starting from the 0 value for groung floor (source: advertiser)}
  \item{FLATLOCATIONID}{Indicates the kind of views the flat has (1 - external, 2 - internal)}
  \item{CADCONSTRUCTIONYEAR}{Construction year as of cadastral source (source: cadastre), note this figure can differ from the one given by the advertiser}
  \item{CADMAXBUILDINGFLOOR}{Max building floor (source: cadastre)}
  \item{CADDWELLINGCOUNT}{Dwelling count in the building (source: cadastre)}
  \item{CADASTRALQUALITYID}{Cadastral quality (source: cadastre)}
  \item{BUILTTYPEID_1}{Dummy value for flat condition: 1 new development 0 otherwise (source: advertiser))}
  \item{BUILTTYPEID_2}{Dummy value for flat condition: 1 second hand to be restored 0 otherwise (source: advertiser))}
  \item{BUILTTYPEID_3}{Dummy value for flat condition: 1 second hand in good condition 0 otherwise (source: advertiser))}
  \item{DISTANCE_TO_CITY_CENTER}{Distance to center of city in km}
  \item{DISTANCE_TO_DIAGONAL}{Distance to Avinguda Diagonal in km; Diagonal is a major street that cuts across the city diagonally to the street grid}
  \item{geometry}{geometry of simple features}
  }

An object of class \code{sf} (inherits from \code{data.frame}) with 61486 rows and 44 columns.
}
\source{
Idealista (https://www.idealista.com/)
}
\usage{
data(Barcelona_Sale)

Barcelona_Sale
}
\description{
A dataset containing information about 84,280 real estate ads on idealista
in Barcelona during 2018.
}
\examples{
 library(idealista18)
 data(Barcelona_Sale)
 summary(Barcelona_Sale$PRICE)
}
\keyword{datasets}
