% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealista18.R
\docType{package}
\name{idealista18}
\alias{idealista18}
\title{idealista18: A data package with property values in Spain from the Idealista data base.}
\description{
This package is an exercise in package creation using
R studio. The package includes a sample function and
a sample dataset with their respective documentation.
}
\references{
\url{https://github.com/paezha/Reproducible-Research-Workflow}
}
\author{
Antonio Paez, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}
}
