% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealista18.R
\docType{data}
\name{Madrid_Polygons}
\alias{Madrid_Polygons}
\title{Neighborhood polygons in Madrid}
\format{
An object of class `sf` (inherits from `data.frame`) with 73 rows and 4 variables:
\describe{
  \item{LOCATIONID}{ID of the neighborhood}
  \item{LOCATIONNAME}{Name of the neighborhood)}
  \item{geometry}{geometry of simple features}
  }

An object of class \code{sf} (inherits from \code{data.frame}) with 135 rows and 4 columns.
}
\source{
Idealista (https://www.idealista.com/)
}
\usage{
data(Madrid_Polygons)

Madrid_Polygons
}
\description{
Neighborhoods in the city of Madrid
}
\examples{
 library(idealista18)
 data(Madrid_Polygons)
 summary(Madrid_Polygons)
}
\keyword{datasets}
