% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealista18.R
\docType{data}
\name{Madrid_Sale}
\alias{Madrid_Sale}
\title{Sale prices of residential property in Madrid in 2018.}
\format{


An object of class \code{sf} (inherits from \code{data.frame}) with 94815 rows and 42 columns.
}
\source{
Idealista (https://www.idealista.com/)
}
\usage{
data(Madrid_Sale)

Madrid_Sale
}
\description{
A dataset containing information about 156,016 real estate listings on idealista
in Madrid during 2018.
}
\examples{
 library(idealista18)
 data(Madrid_Sale)
 summary(Madrid_Sale$PRICE)
}
\keyword{datasets}
