% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf_reg_metrics.R
\name{perf_reg_metrics}
\alias{perf_reg_metrics}
\title{Compute various performance metrics for a regression model}
\usage{
perf_reg_metrics(truth, predictions, model_label = "", split_label = "")
}
\arguments{
\item{truth}{A simple features data frame with the data for model estimation.}

\item{predictions}{The proportion of the sample for training, a number between 0 and 1.}

\item{model_label}{A label for the model}

\item{split_label}{A label for the data split used}
}
\value{
A list with various performance metrics.
}
\description{
Compute various performance metrics for a regression model
}
\examples{

library(idealista18)
data(Valencia_Sale)
mod <- lm(LOGPRICE ~ CONSTRUCTEDAREA, data = Valencia_Sale)
perf_reg_metrics(truth = Valencia_Sale$LOGPRICE, predictions = mod$fitted.values)
}
