% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealista18.R
\docType{data}
\name{properties_by_district}
\alias{properties_by_district}
\title{Total number of properties by district in 2018.}
\format{
An object of class `sf` (inherits from `data.frame`) with 50 rows and 5 variables:
\describe{
  \item{District}{Number of district}
  \item{N_CADASTRE}{Date YYYYMM, indicates the quarter when the ad was extracted we used YYYY03 for the 1st Q, YYYY06 the 2nd, YYYY09 for the 3rd and YYYY12 for the 4th}
  \item{N}{Asking price for the ad at idealista expressed in euros}
  \item{CITY}{Price in euros per square meter}
  \item{geometry}{geometry of simple features in latitude and longitude}
  }

An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 47 rows and 5 columns.
}
\source{
Idealista (https://www.idealista.com/)
}
\usage{
data(properties_by_district)

properties_by_district
}
\description{
A dataset containing information about the number of properties by district in Barcelona, Madrid, and Valencia during 2018 according to the cadastre. Districts are collections of neighborhoods.
}
\examples{
 library(idealista18)
 data(properties_by_district)
 summary(properties_by_district$N)
}
\keyword{datasets}
