% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealista18.R
\docType{data}
\name{Valencia_POIS}
\alias{Valencia_POIS}
\title{Places of interest in Valencia}
\format{
A list with three objects:
\describe{
  \item{City_Center}{Coordinates of Valencia's city center}
  \item{Metro}{Coordinates of Valencia's metro}
  \item{Blasco}{Coordinates of Blasco)}
  }

An object of class \code{list} of length 3.
}
\source{
Idealista (https://www.idealista.com/)
}
\usage{
data(Valencia_POIS)

Valencia_POIS
}
\description{
Geospatial references for places of interest in Valencia
}
\examples{
 library(idealista18)
 data(Valencia_POIS)
 summary(Valencia_POIS)
}
\keyword{datasets}
