% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealista18.R
\docType{package}
\name{idealista18}
\alias{idealista18}
\title{idealista18: A data package with real estate listings in Spain from the idealista data base.}
\description{
This package contains four quarterly extractions with the set of apartments listed on idealista during 2018,
the package contains three data items for the three spanish cities of Madrid, Barcelona and Valencia:
}
\details{
* Polygons for each city neighborhood
* Listing ads (referred also as assets)
* Points of Interest for each city

Escribir algo
}
\references{
\url{https://github.com/paezha/Reproducible-Research-Workflow}
}
\author{
\tabular{ll}{
  David Rey \tab \email{drey@idealista.com} \cr
  Pelayo Arbues \tab \email{pgarbues@idealista.com} \cr
  Fernando López  \tab \email{fernando.lopez@upct.es} \cr
  Antonio Páez \tab \email{paezha@gmail.com} \cr
  }
}
