% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealista18.R
\docType{package}
\name{idealista18}
\alias{idealista18}
\title{idealista18: A data package with real estate litings in Spain from the idealista data base.}
\description{
This package contains four quarterly extractions with the set of apartments listed on idealista during 2018,
the package contains three data items for the three spanish cities of Madrid, Barcelona and Valencia:
}
\details{
* Polygons for each city neighborhood
* Listing ads (referred also as assets)
* Points of Interest for each city
}
\references{
\url{https://github.com/paezha/Reproducible-Research-Workflow}
}
\author{
Antonio Paez, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}
}
